ortoreal/requirements.txt
asgiref==3.6.0
astroid==2.15.1
Babel==2.12.1
cfgv==3.3.1
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
dill==0.3.6
distlib==0.3.6
Django==4.2
django-appconf==1.0.5
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap5==23.1
django-compressor==4.3.1
django-extensions==3.2.1
django-filter==23.2
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-sass-processor==1.2.2
django-select2==8.1.1
django-tables2==2.5.3
djangorestframework==3.14.0
djlint==1.19.16
EditorConfig==0.12.3
et-xmlfile==1.1.0
filelock==3.11.0
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
identify==2.5.22
isort==5.12.0
jsbeautifier==1.14.7
lazy-object-proxy==1.9.0
libsass==0.22.0
mccabe==0.7.0
nodeenv==1.7.0
openpyxl==3.1.2
pathspec==0.11.1
phonenumbers==8.13.9
platformdirs==3.2.0
pre-commit==3.2.2
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
pydantic==1.10.8
pylint==2.17.1
pymemcache==4.0.0
pytz==2023.3
PyYAML==6.0
rcssmin==1.1.1
regex==2022.10.31
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tomlkit==0.11.7
tqdm==4.65.0
typing_extensions==4.6.2
virtualenv==20.21.0
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.1.0
xlwt==1.3.0